# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: |
    Tree Friend E-Commerce API.

    ## Versioning

    The API is mounted under two prefixes:
      - `/api/v1` — the canonical, versioned path. New clients should use this.
      - `/api`     — backward-compat alias for existing callers. Same handlers,
                     transparently aliased.

    When a v2 is needed, a new router will be mounted under `/api/v2` and
    `/api/v1` will be left untouched. The unversioned `/api` alias will be
    deprecated (Deprecation header + warning log) once all callers have
    migrated to `/api/v1`.
servers:
  - url: /api/v1
    description: Versioned API path (canonical)
  - url: /api
    description: Unversioned API path (backward-compat alias)
...
paths:
...
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /bkash/create-payment:
    post:
      operationId: createBkashPayment
      tags: [bkashPayment]
      summary: Start a bKash Tokenized Checkout payment for an already-created order (authenticated buyer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBkashPaymentBody"
      responses:
        "200":
          description: bKash payment created — redirect the buyer to bkashURL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BkashPaymentSession"
  /bkash/create-payment/guest:
    post:
      operationId: createBkashPaymentGuest
      tags: [bkashPayment]
      summary: Start a bKash Tokenized Checkout payment for an already-created guest order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBkashPaymentGuestBody"
      responses:
        "200":
          description: bKash payment created — redirect the buyer to bkashURL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BkashPaymentSession"
  /bkash/query-payment/{paymentID}:
    get:
      operationId: queryBkashPayment
      tags: [bkashPayment]
      summary: Admin — look up a bKash payment's live status directly from bKash (reconciliation/debugging)
      parameters:
        - name: paymentID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: bKash's own payment status for this paymentID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BkashQueryPaymentResult"
...
components:
  schemas:
...
    OrderLineItem:
      type: object
      description: >-
        Snapshot of a purchased line item, stored as JSON on the order at
        checkout time. Either variantId+variantName are set (admin-direct
        line) or sellerListingId+sellerId are set (marketplace line) --
        never both. Every line on a given order has the same kind and, for
        marketplace orders, the same sellerId as the order itself.
      properties:
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: string
        variantId:
          type: integer
        variantName:
          type: string
        sellerListingId:
          type: integer
        sellerId:
          type: integer
        quantity:
          type: integer
        price:
          type: number
        deliveryCharge:
          type: number
      required:
        - productId
        - productName
        - productImage
        - quantity
        - price
        - deliveryCharge
    Order:
      type: object
      properties:
        id:
          type: integer
        trackingId:
          type: string
        userId:
          type: string
        sellerId:
          type: ["integer", "null"]
          description: Null for admin-direct orders; set for marketplace orders.
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderLineItem"
        totalAmount:
          type: number
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        senderNumber:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
        orderStatus:
          type: string
        transactionId:
          type: ["string", "null"]
        shippingAddress:
          $ref: "#/components/schemas/AddressBody"
        couponCode:
          type: ["string", "null"]
        discountAmount:
          type: number
        cancellationReason:
          type: ["string", "null"]
        giftWrap:
          type: string
        giftMessage:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - trackingId
        - userId
        - items
        - totalAmount
        - paymentMethod
        - paymentStatus
        - orderStatus
        - transactionId
        - shippingAddress
        - couponCode
        - discountAmount
        - createdAt
        - updatedAt
...
    CreateBkashPaymentBody:
      type: object
      description: >-
        Authenticated path — orderId must belong to the calling user, be
        paymentMethod "bkash", and currently be paymentStatus
        "payment_pending" (i.e. not already paid/failed/refunded).
      properties:
        orderId:
          type: integer
      required:
        - orderId
    CreateBkashPaymentGuestBody:
      type: object
      description: >-
        Guest path — trackingId scopes to a guest order (userId starting
        with "guest_") only; the tracking id itself is the bearer secret,
        same trust model as GET /orders/track/{trackingId}.
      properties:
        trackingId:
          type: string
      required:
        - trackingId
    BkashPaymentSession:
      type: object
      description: >-
        Redirect the buyer's browser (or open a popup/webview) to bkashURL
        to complete authorization on bKash's own hosted page. bKash will
        redirect back to our server's own /bkash/callback afterward, which
        then redirects the buyer on to the order detail page with a
        ?bkash=... status flag.
      properties:
        paymentID:
          type: string
        bkashURL:
          type: string
        orderId:
          type: integer
        trackingId:
          type: string
      required:
        - paymentID
        - bkashURL
        - orderId
        - trackingId
    BkashQueryPaymentResult:
      type: object
      description: Raw-ish status straight from bKash's own Query Payment API, for reconciliation/debugging only.
      properties:
        paymentID:
          type: string
        trxID:
          type: ["string", "null"]
        transactionStatus:
          type: string
        amount:
          type: string
        merchantInvoiceNumber:
          type: ["string", "null"]
          description: Our own orders.trackingId, as originally sent to bKash at Create Payment time.
      required:
        - paymentID
        - transactionStatus
...
    AddressBody:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        street:
          type: string
        city:
          type: string
        district:
          type: string
        postalCode:
          type: ["string", "null"]
        isDefault:
          type: boolean
      required:
        - fullName
        - phone
        - street
        - city
        - district
...